db: use uint64 for batch's memTableSize
A memtable node has more overhead than a batch's wire encoding, which
means a batch's memtable size grows faster than the batch's backing byte
slice. We enforce a maxBatchSize on size of the batch's backing byte
slice. Previously, if a batch approached the maxBatchSize but didn't
cross it, the memTableSize, a uint32, could overflow. This overflow
would cause Pebble to incorrectly add the batch to the memtable rather
than the flush list and to reserve the wrong amount of space in the
memtable.

This commit fixes the issue by tracking the batch's memTableSize using a
uint64.

See cockroachdb/cockroach#58004.

+++ batch.go
@@ -187,7 +187,12 @@ type Batch struct {
 	formatKey      base.FormatKey
 	abbreviatedKey AbbreviatedKey
 
-	memTableSize uint32
+	// The required space to add this batch to a memtable.
+	// Note that although batches are limited to 4 GiB in size, that limit
+	// applies to len(data), not the memtable size. A memtable node adds more
+	// overhead than the batch's log encoding, so memTableSize is larger than
+	// len(data) and may overflow a uint32.
+	memTableSize uint64
 
 	// The db to which the batch will be committed. Do not change this field
 	// after the batch has been created as it might invalidate internal state.
